Issue 816: Pull from master

+++ src/common/operator/cast_operators.cpp
@@ -32,7 +32,7 @@ template <class SRC, class DST> static bool try_cast_with_overflow_check(SRC val
 template <class SRC, class DST> static DST cast_with_overflow_check(SRC value) {
 	DST result;
 	if (!try_cast_with_overflow_check<SRC, DST>(value, result)) {
-		throw ValueOutOfRangeException((int64_t)value, GetTypeId<SRC>(), GetTypeId<DST>());
+		throw ValueOutOfRangeException((double)value, GetTypeId<SRC>(), GetTypeId<DST>());
 	}
 	return result;
 }
@@ -1264,15 +1264,15 @@ template <> int16_t CastToDecimal::Operation(bool input, uint8_t width, uint8_t 
 }
 
 template <> int32_t CastToDecimal::Operation(bool input, uint8_t width, uint8_t scale) {
-    return input ? NumericHelper::PowersOfTen[scale] : 0;
+	return input ? NumericHelper::PowersOfTen[scale] : 0;
 }
 
 template <> int64_t CastToDecimal::Operation(bool input, uint8_t width, uint8_t scale) {
-    return input ? NumericHelper::PowersOfTen[scale] : 0;
+	return input ? NumericHelper::PowersOfTen[scale] : 0;
 }
 
 template <> hugeint_t CastToDecimal::Operation(bool input, uint8_t width, uint8_t scale) {
-    return input ? Hugeint::PowersOfTen[scale] : 0;
+	return input ? Hugeint::PowersOfTen[scale] : 0;
 }
 
 template <> bool CastFromDecimal::Operation(int16_t input, uint8_t width, uint8_t scale) {
@@ -1308,7 +1308,7 @@ template <class SRC> hugeint_t NumericToHugeDecimalCast(SRC input, uint8_t width
 	hugeint_t max_width = Hugeint::PowersOfTen[width - scale];
 	hugeint_t hinput = hugeint_t(input);
 	if (hinput >= max_width || hinput <= -max_width) {
-		throw OutOfRangeException("Could not cast value %s to DECIMAL(%d,%d)", hinput.ToString() , width, scale);
+		throw OutOfRangeException("Could not cast value %s to DECIMAL(%d,%d)", hinput.ToString(), width, scale);
 	}
 	return hinput * Hugeint::PowersOfTen[scale];
 }
